robot_interaction: split InteractionHandler into its own file
